revert: revert renaming project to entity
