+++ src/main/java/org/discordbots/api/client/entity/Result.java
@@ -1,4 +1,4 @@
-package org.discordbots.api.client.project;
+package org.discordbots.api.client.entity;
 
 import java.util.List;
