+++ src/main/java/org/discordbots/api/client/entity/Vote.java
@@ -1,4 +1,4 @@
-package org.discordbots.api.client.project;
+package org.discordbots.api.client.entity;
 
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
