+++ src/main/java/org/discordbots/api/client/entity/VotingMultiplier.java
@@ -1,4 +1,4 @@
-package org.discordbots.api.client.project;
+package org.discordbots.api.client.entity;
 
 import com.google.gson.annotations.SerializedName;
